+++ CHANGELOG.md
@@ -18,6 +18,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Update `.gitignore`
 - Move `astar.ts` to `lib`
 
+### Fixed
+- Bug that empty array could go into the main while loop of the astar-finder
+
 ## [1.0.0] - 2019-07-26
 ### Added
 - Example folder with an integration of this library into [Phaser 3](https://phaser.io/phaser3)